Support fetching update config from remote server [1/3]
This change introduces support for fetching update config from our
remote server while preserving the ability to use local config.

Developers can push a local config to the app's file directory under
device's codename and the app will automatically pick it up. This is
helpful to test a local config. This will be removed in future. We will
switch to mock web server and have a test instead.

This patch also adds the following libraries for convenience:
- Dagger-Hilt for dependency injection
- ktlint for formatting kotlin code

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>
Change-Id: I835bb5803d252f2a2ed045f0ab0c8c74a898f637