Move status & config  handling to UpdateManager
This results in UI being dumb and UpdateManager SSOT. This is
required as we would like automated update check in background.

This change also saves the JSON config locally as a file for
refetching updates. This file is invalidated after 3 hours have
passed and a new JSON is fetched again aligning with auto-updates
check.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>
Change-Id: I071e0b9c293435b48e57786a2ef666100581a15e